[FIX] Directory Update module
When using the directory update module there might be an error if the diskspace does not suffice to extract the provided artifact.

The return value of the extraction is not checked and the script keeps execution till it returns 0. So though the destination directory might be incompletely extracted, the script reports an install success.

This commit fixes the issue by explicitly checking for the return value of the compress and extract command.